account_info serialize and unit test

+++ logos/consensus/persistence/batchblock/batchblock_persistence.hpp
@@ -55,7 +55,7 @@ private:
                       uint64_t timestamp,
                       MDB_txn * transaction);
 
-    static constexpr uint64_t  RESERVATION_PERIOD  = 2;
+    static constexpr uint32_t  RESERVATION_PERIOD  = 2;
     static constexpr uint128_t MIN_TRANSACTION_FEE = 0x21e19e0c9bab2400000_cppui128; // 10^22
 
     Log                 _log;
